Fix sockopts

+++ sockopts.go
@@ -20,10 +20,10 @@ func (s *Socket) GetType() SocketType {
 	return SocketType(s.getInt(C.ZMQ_TYPE))
 }
 func (s *Socket) GetSendHWM() uint64 {
-	return uint64(s.getInt64(C.ZMQ_SNDHWM))
+	return uint64(s.getInt(C.ZMQ_SNDHWM))
 }
 func (s *Socket) GetRecvHWM() uint64 {
-	return uint64(s.getInt64(C.ZMQ_RCVHWM))
+	return uint64(s.getInt(C.ZMQ_RCVHWM))
 }
 func (s *Socket) GetRecvTimeout() time.Duration {
 	return toDuration(int64(s.getInt(C.ZMQ_RCVTIMEO)), time.Millisecond)
@@ -73,9 +73,9 @@ func (s *Socket) GetEvents() EventSet {
 
 /* Set */
 
-func (s *Socket) SetHWM(hwm uint64) {
-	s.setInt64(C.ZMQ_SNDHWM, int64(hwm))
-	s.setInt64(C.ZMQ_RCVHWM, int64(hwm))
+func (s *Socket) SetHWM(hwm int) {
+	s.setInt(C.ZMQ_SNDHWM, int(hwm))
+	s.setInt(C.ZMQ_RCVHWM, int(hwm))
 }
 func (s *Socket) SetAffinity(affinity uint64) {
 	s.setInt64(C.ZMQ_AFFINITY, int64(affinity))
